refactor(assistant): move channel stores out of memory/ into channels/
