+++ assistant/src/channels/channel-verification-sessions.ts
