+++ assistant/src/channels/slack-thread-store.ts
