+++ assistant/src/runtime/channel-verification-service.ts
@@ -9,24 +9,13 @@
 import { createHash, randomBytes } from "crypto";
 import { v4 as uuid } from "uuid";
 
-import { revokeGuardianBinding } from "../contacts/contacts-write.js";
-import {
-  getGuardianDelivery,
-  getGuardianDeliveryFresh,
-  guardianForChannel,
-} from "../contacts/guardian-delivery-reader.js";
-import {
-  getRateLimit,
-  recordInvalidAttempt,
-  resetRateLimit,
-} from "../contacts/guardian-rate-limits.js";
 import type {
   GuardianBinding,
   IdentityBindingStatus,
   SessionStatus,
   VerificationPurpose,
   VerificationSession,
-} from "../memory/channel-verification-sessions.js";
+} from "../channels/channel-verification-sessions.js";
 import {
   bindSessionIdentity as storeBindSessionIdentity,
   consumeSession,
@@ -41,7 +30,18 @@ import {
   revokePendingSessions as storeRevokePendingSessions,
   updateSessionDelivery as storeUpdateSessionDelivery,
   updateSessionStatus as storeUpdateSessionStatus,
-} from "../memory/channel-verification-sessions.js";
+} from "../channels/channel-verification-sessions.js";
+import { revokeGuardianBinding } from "../contacts/contacts-write.js";
+import {
+  getGuardianDelivery,
+  getGuardianDeliveryFresh,
+  guardianForChannel,
+} from "../contacts/guardian-delivery-reader.js";
+import {
+  getRateLimit,
+  recordInvalidAttempt,
+  resetRateLimit,
+} from "../contacts/guardian-rate-limits.js";
 import { composeApprovalMessage } from "./approval-message-composer.js";
 
 // ---------------------------------------------------------------------------
